requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pandas
numpy
matplotlib
plotly
jinja2
weasyprint
reportlab
yfinance
slowapi
fastapi-users
authlib
python-dotenv
starlette
httpx
pytest
pytest-asyncio
coverage
pandas-ta
finnhub-python
redis
websocket